add: date functions file
